@@ -28,31 +28,31 @@
     "clean": "rimraf dist"
   },
   "peerDependencies": {
-    "react-admin": "3.x"
+    "react-admin": "4.x"
   },
   "dependencies": {
     "papaparse": "5.x",
-    "ramda": "^0.27.0"
+    "ramda": "^0.28.0",
+    "react-redux": "^8.0.1"
   },
   "devDependencies": {
-    "@types/jest": "24.0.11",
-    "@types/node": "11.12.0",
-    "@types/papaparse": "4.x",
-    "@types/ramda": "0.27.34",
-    "@types/react": "16.8.10",
-    "@types/react-dom": "16.8.3",
-    "@types/react-redux": "^7.1.7",
+    "@types/jest": "27.4.1",
+    "@types/node": "17.0.25",
+    "@types/papaparse": "5.x",
+    "@types/ramda": "0.28.11",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
     "jest-coverage-badges": "^1.1.2",
-    "ra-data-fakerest": "^3.2.2",
-    "ra-language-chinese": "^2.0.5",
+    "ra-data-fakerest": "^4.0.1",
+    "ra-language-chinese": "^2.0.10",
     "ra-language-russian": "^3.13.5",
     "ra-language-spanish": "1.x",
-    "react": "16.8.6",
-    "react-admin": "3.10.x",
-    "react-dom": "16.8.3",
-    "react-scripts": "^3.4.3",
-    "source-map-loader": "^0.2.4",
-    "ts-jest": "^24.x.x",
-    "typescript": "3.x"
+    "react": "^17.0.0",
+    "react-admin": "4.0.x",
+    "react-dom": "^17.0.0",
+    "react-scripts": "^5.0.1",
+    "source-map-loader": "^3.0.1",
+    "ts-jest": "^27.x.x",
+    "typescript": "4.x"
   }
 }
