@@ -184,7 +184,7 @@ export const MainCsvImport = (props: any) => {
   const handleClose = () => {
     console.log("handleClose", { file });
     resetVars();
-    notify(translate("csv.dialogImport.alertClose", { fname: fileName }), 'info');
+    notify(translate("csv.dialogImport.alertClose", { fname: fileName }), { type: 'info'});
     refresh();
   };
 
